Rename ScanQR to SendByQR and use it only for sending
- Renamed to SendByQR
- Removed Receive flow
